api/requirements.txt
astroid==2.2.4
autopep8==1.4.3
appnope==0.1.0
Babel==2.6.0
backcall==0.1.0
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.3.0
dj-database-url==0.5.0
Django==2.1.7
django-environ==0.4.5
django-extensions==2.1.3
django-filter==2.0.0
django-filters==0.2.1
django-heroku==0.3.1
django-macros==0.4.0
django-mail-templated==2.6.5
django-phonenumber-field==2.0.1
django-redis==4.9.0
django-rest-swagger==2.2.0
django-settings-export==1.2.1
django-templated-mail==1.1.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.0.0
djoser==1.3.0
environ==1.0
gunicorn==19.9.0
idna==2.7
inflection==0.3.1
ipdb==0.11
ipython==7.0.1
ipython-genutils==0.2.0
isort==4.3.10
itypes==1.1.0
jedi==0.13.1
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
openapi-codec==1.3.2
parso==0.3.1
pep8==1.7.1
pexpect==4.6.0
phonenumbers==8.9.15
pickleshare==0.7.5
pipenv==2018.11.26
prompt-toolkit==2.0.5
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.2.0
PyJWT==1.7.1
pylint==2.3.1
pytz==2018.5
redis==2.10.6
requests==2.20.0
simplegeneric==0.8.1
simplejson==3.16.0
six==1.11.0
traitlets==4.3.2
typed-ast==1.3.1
uritemplate==3.0.0
urllib3==1.23
virtualenv==16.4.3
virtualenv-clone==0.5.1
wcwidth==0.1.7
whitenoise==4.1.2
wrapt==1.11.1
